Move the Wasm-compatible `Image` from `static` to `shared`
