+++ js/image/static/index.ts
@@ -1,2 +1,1 @@
 export { default } from "./StaticImage.svelte";
-export { default as Image } from "./Image.svelte";
